LPS-111943 Extract to util file to reduce file length
