+++ modules/dxp/apps/analytics/analytics-reports-web/src/main/resources/META-INF/resources/js/components/Chart.js
@@ -27,6 +27,7 @@ import {
 
 import ConnectionContext from '../state/context';
 import {useChartState} from '../utils/chartState';
+import {generateDateFormatters as dateFormat} from '../utils/dateFormat';
 import {numberFormat} from '../utils/numberFormat';
 import {ActiveDot as CustomActiveDot, Dot as CustomDot} from './CustomDots';
 import CustomTooltip from './CustomTooltip';
@@ -99,91 +100,6 @@ function thousandsToKilosFormater(value) {
 
 	return value;
 }
-
-/*
- * It generates a set of functions used to produce
- * internationalized date related content.
- */
-const generateDateFormatters = key => {
-	/*
-	 * Given 2 date objects it produces a user friendly date interval
-	 *
-	 * For 'en-US'
-	 * [Date, Date] => '16 - Jun 21, 2020'
-	 */
-	function formatChartTitle([initialDate, finalDate]) {
-		const singleDayDateRange =
-			finalDate - initialDate <= 1000 * 60 * 60 * 24;
-
-		const dateFormatter = (
-			date,
-			options = {
-				day: 'numeric',
-				month: 'short',
-				year: 'numeric',
-			}
-		) => Intl.DateTimeFormat([key], options).format(date);
-
-		const equalMonth = initialDate.getMonth() === finalDate.getMonth();
-		const equalYear = initialDate.getYear() === finalDate.getYear();
-
-		const initialDateOptions = {
-			day: 'numeric',
-			month: equalMonth && equalYear ? undefined : 'short',
-			year: equalYear ? undefined : 'numeric',
-		};
-
-		if (singleDayDateRange) {
-			return dateFormatter(finalDate);
-		}
-
-		return `${dateFormatter(
-			initialDate,
-			initialDateOptions
-		)} - ${dateFormatter(finalDate)}`;
-	}
-
-	/*
-	 * Given a date like string it produces a internationalized long date
-	 *
-	 * For 'en-US'
-	 * String => '06/17/2020'
-	 */
-	function formatLongDate(value) {
-		return Intl.DateTimeFormat([key]).format(new Date(value));
-	}
-
-	/*
-	 * Given a date like string produces the day of the month
-	 *
-	 * For 'en-US'
-	 * String => '16'
-	 */
-	function formatNumericDay(value) {
-		return Intl.DateTimeFormat([key], {
-			day: 'numeric',
-		}).format(new Date(value));
-	}
-
-	/*
-	 * Given a date like string produces the hour of the day
-	 *
-	 * For 'en-US'
-	 * String => '04 AM'
-	 */
-	function formatNumericHour(value) {
-		return Intl.DateTimeFormat([key], {
-			hour: 'numeric',
-		}).format(new Date(value));
-	}
-
-	return {
-		formatChartTitle,
-		formatLongDate,
-		formatNumericDay,
-		formatNumericHour,
-	};
-};
 
 function legendFormatterGenerator(
 	totals,
@@ -289,7 +205,7 @@ export default function Chart({
 		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [chartState.timeSpanOption, chartState.timeSpanOffset]);
 
-	const dateFormatters = useMemo(() => generateDateFormatters(languageTag), [
+	const dateFormatters = useMemo(() => dateFormat(languageTag), [
 		languageTag,
 	]);
